fix deprecated APIs and lint warnings
