+++ src/lib/toast-context.ts
@@ -1,0 +1,23 @@
+import { createContext, useContext } from "react";
+
+export type ToastVariant = "default" | "success" | "error" | "warning";
+
+export interface Toast {
+  id: string;
+  message: string;
+  variant: ToastVariant;
+}
+
+export interface ToastContextValue {
+  toasts: Toast[];
+  toast: (message: string, variant?: ToastVariant) => void;
+  dismiss: (id: string) => void;
+}
+
+export const ToastContext = createContext<ToastContextValue | null>(null);
+
+export const useToast = () => {
+  const ctx = useContext(ToastContext);
+  if (!ctx) throw new Error("useToast must be used within ToastProvider");
+  return ctx;
+};
