+++ src/lib/toast.tsx
@@ -1,26 +1,6 @@
-import { createContext, type ReactNode, useCallback, useContext, useState } from "react";
+import { type ReactNode, useCallback, useState } from "react";
 
-export type ToastVariant = "default" | "success" | "error" | "warning";
-
-interface Toast {
-  id: string;
-  message: string;
-  variant: ToastVariant;
-}
-
-interface ToastContextValue {
-  toasts: Toast[];
-  toast: (message: string, variant?: ToastVariant) => void;
-  dismiss: (id: string) => void;
-}
-
-const ToastContext = createContext<ToastContextValue | null>(null);
-
-export const useToast = () => {
-  const ctx = useContext(ToastContext);
-  if (!ctx) throw new Error("useToast must be used within ToastProvider");
-  return ctx;
-};
+import { type Toast, ToastContext, type ToastVariant } from "@/lib/toast-context";
 
 export const ToastProvider = ({ children }: { children: ReactNode }) => {
   const [toasts, setToasts] = useState<Toast[]>([]);
