+++ src/routes/_auth.settings.tsx
@@ -9,7 +9,7 @@ import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { useUpdateUserSettings, useUserSettings } from "@/hooks/useUserSettings";
-import { useToast } from "@/lib/toast";
+import { useToast } from "@/lib/toast-context";
 
 const SettingsPage = () => {
   const { t } = useTranslation("settings");
