+++ vite.config.ts
@@ -1,7 +1,7 @@
 import path from "path";
 
 import tailwindcss from "@tailwindcss/vite";
-import { TanStackRouterVite } from "@tanstack/router-plugin/vite";
+import { tanstackRouter } from "@tanstack/router-plugin/vite";
 import react from "@vitejs/plugin-react";
 import { defineConfig } from "vite";
 import { VitePWA } from "vite-plugin-pwa";
@@ -11,7 +11,7 @@ const isStorybook = process.env.STORYBOOK === "true";
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [
-    TanStackRouterVite({
+    tanstackRouter({
       routesDirectory: "./src/routes",
       generatedRouteTree: "./src/routeTree.gen.ts",
     }),
@@ -44,6 +44,9 @@ export default defineConfig({
             },
             devOptions: {
               enabled: false,
+            },
+            injectManifest: {
+              rollupFormat: "iife",
             },
           }),
         ]),
